latest changes

+++ ClassifyImage.py
@@ -8,8 +8,12 @@ imagePath = '/Users/srlaxminaarayanan/Pictures/images'
 
 
 def ProcessResult(imagepath, result):
-    with open("Output.txt", "w+") as text_file:
-        text_file.write("file: %s \n classification: \n %s" % (imagepath, ', '.join(result)))
+    if(os.path.exists("Output.txt")):
+        with open("Output.txt", "a") as text_file:
+            text_file.write("\n file: %s \n classification: \n %s \n" % (imagepath, ', '.join(result)))
+    else:
+        with open("Output.txt", "w+") as text_file:
+            text_file.write("file: %s \n classification: \n %s" % (imagepath, ', '.join(result)))
 
 
 def ClassifyImages():
